Remove redundant 'PermissionsEditor' page
